First wave of updates
Added in more functional stalactites, and the ability for possessed objects to hand off control to other scripts. Bettered the functionality of the GameManger instance, so it's not a full on singleton, rather than just pretending to be. This means that you no longer to remember to put it in the level, just calling Gamemanger.instance is enough to spawn it and have it set everything up for you. Tweaked possessable objects to have better control, and added the ability for objects to deal damage. Damaging objects can also have a set speed they need to reach before they damage anything. Gizmos got updated a little bit, so now there are speed indicators that line up with possessable objects. Possessable objects you should be able to jump on are in a new layer, configured so that the player raycasts work with it. Player collider was fixed to stop the wall issue, minor bug there but it'll be fixed with the next update.